pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "abhishek-yanjarappa", email = "75441358+abhishek-yanjarappa@users.noreply.github.com"},
]
dependencies = [
    "gunicorn>=20.1.0",
    "python-telegram-bot>=13.13",
    "quart>=0.17.0",
    "python-dotenv>=0.20.0",
    "pymongo>=4.1.1",
]
requires-python = ">=3.10"
license = {text = "MIT"}

[build-system]
requires = ["pdm-pep517>=1.0.0"]
build-backend = "pdm.pep517.api"

[tool]
[tool.pdm]

[tool.pdm.scripts]
dev.cmd = "quart run"
dev.env = {QUART_APP = "main:app", QUART_ENV = "development"}
